+++ CHANGELOG.md
@@ -12,6 +12,11 @@ A preview of the next release can be installed from
 - Bump babashka.http-client to v0.4.15
 - Bump rewrite-clj to v0.1.1.47
 - [#1619](https://github.com/babashka/babashka/issues/1619): Fix reflection issue with `Thread/sleep` in `core.async/timeout`
+- Support interop on `java.util.stream.IntStream`
+- [#1513](https://github.com/babashka/babashka/issues/1513): Fix interop on `Thread/sleep` with numbers that aren't already longs
+- [#1624](https://github.com/babashka/babashka/pull/1624): Use Oracle GraalVM 21 ([@lispyclouds](https://github.com/lispyclouds))
+- Use PGO to speed up loops (now 2-3x faster for `(time (loop [val 0 cnt 10000000] (if (pos? cnt) (recur (inc val) (dec cnt)) val)))`!)
+- Bump babashka.cli to 0.7.53
 
 ## 1.3.184 (2023-08-22)